requirements.txt
apache-airflow>=2.0
flake8>=5.0
pytest~=7.1
pytest-cov~=4.0
